--- dawstore-app/resources/js/app.js
@@ -1,1 +1,0 @@
-import './bootstrap';
